tests: add tests
